+++ src/smp/features/categorical.py
@@ -1,5 +1,5 @@
 import numpy as np
-from sklearn.impute import SimpleImputer, MissingIndicator
+from sklearn.impute import SimpleImputer
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import OneHotEncoder, LabelBinarizer
 from smp.features.features import Feature, Base, ToVector
@@ -17,15 +17,6 @@ class LowerCase(Base):
 class Categorical(Feature):
     def __init__(self, var_name):
         super().__init__(var_name)
-
-
-class PersonalURL(Categorical):
-    def __init__(self):
-        super().__init__("Personal URL")
-        self._pipe = Pipeline(
-            [ToVector().to_step(), ("MissingIndicator", MissingIndicator()),],
-            verbose=True,
-        )
 
 
 class ProfileCoverImageStatus(Categorical):
@@ -51,12 +42,6 @@ class ProfileVerificationStatus(Categorical):
             [ToVector().to_step(), ("OneHotEncoder", OneHotEncoder(drop="first"),),],
             verbose=True,
         )
-
-
-class IsProfileViewSizeCustomized(Categorical):
-    def __init__(self):
-        super().__init__("Is Profile View Size Customized?")
-        self._pipe = Pipeline([ToVector().to_step(),], verbose=True,)
 
 
 class LocationPublicVisibility(Categorical):
